services/image-classification/requirements.txt
Flask==1.1.4
keras==2.7.0
numpy==1.22.0
Pillow==8.4.0
psycopg2_binary==2.9.2
waitress==2.0.0
